refactor: rename prompt experiment artifact helper
Name the helper for the temporary runtime artifacts it manages and update active imports and rerun instructions.
