+++ scripts/prompt-experiments/hyphen-digit/results/20260911-semantic-rule/REPORT.md
@@ -186,7 +186,7 @@ node --input-type=module <<'JS'
 import assert from 'node:assert/strict';
 import { execFileSync } from 'node:child_process';
 import { readFileSync, writeFileSync } from 'node:fs';
-import { outputDirectory } from './scripts/prompt-experiments/public-artifacts.mjs';
+import { outputDirectory } from './scripts/prompt-experiments/artifacts.mjs';
 import { DIAGNOSTIC_QUESTIONS } from './scripts/prompt-experiments/hyphen-digit/results/20260911-semantic-rule/prompts.mjs';
 const report = readFileSync('scripts/prompt-experiments/hyphen-digit/results/20260911-semantic-rule/REPORT.md', 'utf8');
 const selection = JSON.parse(report.match(/```json\n([\s\S]*?)\n```/)[1]);
